Struct wasmer_engine_jit::code_memory::CodeMemory

pub struct CodeMemory {
    unwind_registry: UnwindRegistry,
    mmap: Mmap,
    start_of_nonexecutable_pages: usize,
}

Memory manager for executable code.

Fields
unwind_registry: UnwindRegistrymmap: Mmapstart_of_nonexecutable_pages: usize
Implementations

impl CodeMemory

pub fn new() -> Self

Create a new CodeMemory instance.

pub fn unwind_registry_mut(&mut self) -> &mut UnwindRegistry

Mutably get the UnwindRegistry.

pub fn allocate(
&mut self,
functions: &[&FunctionBody],
executable_sections: &[&CustomSection],
data_sections: &[&CustomSection]
) -> Result<(Vec<&mut [VMFunctionBody]>, Vec<&mut [u8]>, Vec<&mut [u8]>), String>

Allocate a single contiguous block of memory for the functions and custom sections, and copy the data in place.

pub fn publish(&mut self)

Apply the page permissions.

fn function_allocation_size(func: &FunctionBody) -> usize

Calculates the allocation size of the given compiled function.

fn copy_function<'a>(
registry: &mut UnwindRegistry,
func: &FunctionBody,
buf: &'a mut [u8]
) -> &'a mut [VMFunctionBody]

Copies the data of the compiled function to the given buffer.

This will also add the function to the current function table.

fn view_as_mut_vmfunc_slice(slice: &mut [u8]) -> &mut [VMFunctionBody]

Convert mut a slice from u8 to VMFunctionBody.